test/requirements.txt
streamlit
pandas
requests
python-dotenv
psycopg2-binary
